Create scriptSwapEXP.js
this script allows you to swap out whatever script is currently running on purchased servers, and replace it with a file of your choice. It's currently set to attack a list of high value targets with the threads spread evenly across each 1PB server, with a 60s delay from server to server. Not the most efficient but getting better...  Call this version 1.2